Convert float to integers in timer start period

+++ pyqtgraph/SignalProxy.py
@@ -58,7 +58,7 @@ class SignalProxy(QtCore.QObject):
         self.args = args
         if self.rateLimit == 0:
             self.timer.stop()
-            self.timer.start((self.delay * 1000) + 1)
+            self.timer.start(int(self.delay * 1000) + 1)
         else:
             now = time()
             if self.lastFlushTime is None:
@@ -68,7 +68,7 @@ class SignalProxy(QtCore.QObject):
                 leakTime = max(0, (lastFlush + (1.0 / self.rateLimit)) - now)
 
             self.timer.stop()
-            self.timer.start((min(leakTime, self.delay) * 1000) + 1)
+            self.timer.start(int(min(leakTime, self.delay) * 1000) + 1)
 
     def flush(self):
         """If there is a signal queued up, send it now."""
